Fixed Linux build.
